+++ src/util/paymentTypeIcons.ts
@@ -9,6 +9,7 @@ const paymentTypeToThemeKey: Record<FiatPaymentType, keyof Theme | null> = {
   applepay: 'paymentTypeLogoApplePay',
   colombiabank: 'paymentTypeLogoBankTransfer', // Using bank transfer as fallback
   cash: 'paymentTypeLogoCash',
+  cashapp: 'paymentTypeLogoCashApp',
   credit: 'paymentTypeLogoCreditCard',
   directtobank: 'paymentTypeLogoBankTransfer', // Using bank transfer as fallback
   fasterpayments: 'paymentTypeLogoFasterPayments',